feat: ChessBoard 생성 팩토리 분리
